Mark refs dump/restore current api as internal
