+++ clients/python/lakefs_client/api/refs_api.py
@@ -28,7 +28,6 @@ from lakefs_client.model.error import Error
 from lakefs_client.model.find_merge_base_result import FindMergeBaseResult
 from lakefs_client.model.merge import Merge
 from lakefs_client.model.merge_result import MergeResult
-from lakefs_client.model.refs_dump import RefsDump
 
 
 class RefsApi(object):
@@ -134,61 +133,6 @@ class RefsApi(object):
                     'prefix': 'query',
                     'delimiter': 'query',
                     'type': 'query',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [],
-            },
-            api_client=api_client
-        )
-        self.dump_refs_endpoint = _Endpoint(
-            settings={
-                'response_type': (RefsDump,),
-                'auth': [
-                    'basic_auth',
-                    'cookie_auth',
-                    'jwt_token',
-                    'oidc_auth',
-                    'saml_auth'
-                ],
-                'endpoint_path': '/repositories/{repository}/refs/dump',
-                'operation_id': 'dump_refs',
-                'http_method': 'PUT',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'repository',
-                ],
-                'required': [
-                    'repository',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'repository':
-                        (str,),
-                },
-                'attribute_map': {
-                    'repository': 'repository',
-                },
-                'location_map': {
-                    'repository': 'path',
                 },
                 'collection_format_map': {
                 }
@@ -440,68 +384,6 @@ class RefsApi(object):
             },
             api_client=api_client
         )
-        self.restore_refs_endpoint = _Endpoint(
-            settings={
-                'response_type': None,
-                'auth': [
-                    'basic_auth',
-                    'cookie_auth',
-                    'jwt_token',
-                    'oidc_auth',
-                    'saml_auth'
-                ],
-                'endpoint_path': '/repositories/{repository}/refs/restore',
-                'operation_id': 'restore_refs',
-                'http_method': 'PUT',
-                'servers': None,
-            },
-            params_map={
-                'all': [
-                    'repository',
-                    'refs_dump',
-                ],
-                'required': [
-                    'repository',
-                    'refs_dump',
-                ],
-                'nullable': [
-                ],
-                'enum': [
-                ],
-                'validation': [
-                ]
-            },
-            root_map={
-                'validations': {
-                },
-                'allowed_values': {
-                },
-                'openapi_types': {
-                    'repository':
-                        (str,),
-                    'refs_dump':
-                        (RefsDump,),
-                },
-                'attribute_map': {
-                    'repository': 'repository',
-                },
-                'location_map': {
-                    'repository': 'path',
-                    'refs_dump': 'body',
-                },
-                'collection_format_map': {
-                }
-            },
-            headers_map={
-                'accept': [
-                    'application/json'
-                ],
-                'content_type': [
-                    'application/json'
-                ]
-            },
-            api_client=api_client
-        )
 
     def diff_refs(
         self,
@@ -580,71 +462,6 @@ class RefsApi(object):
         kwargs['right_ref'] = \
             right_ref
         return self.diff_refs_endpoint.call_with_http_info(**kwargs)
-
-    def dump_refs(
-        self,
-        repository,
-        **kwargs
-    ):
-        """Dump repository refs (tags, commits, branches) to object store  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.dump_refs(repository, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            repository (str):
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            RefsDump
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['repository'] = \
-            repository
-        return self.dump_refs_endpoint.call_with_http_info(**kwargs)
 
     def find_merge_base(
         self,
@@ -867,73 +684,4 @@ class RefsApi(object):
         kwargs['destination_branch'] = \
             destination_branch
         return self.merge_into_branch_endpoint.call_with_http_info(**kwargs)
-
-    def restore_refs(
-        self,
-        repository,
-        refs_dump,
-        **kwargs
-    ):
-        """Restore repository refs (tags, commits, branches) from object store  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-
-        >>> thread = api.restore_refs(repository, refs_dump, async_req=True)
-        >>> result = thread.get()
-
-        Args:
-            repository (str):
-            refs_dump (RefsDump):
-
-        Keyword Args:
-            _return_http_data_only (bool): response data without head status
-                code and headers. Default is True.
-            _preload_content (bool): if False, the urllib3.HTTPResponse object
-                will be returned without reading/decoding response data.
-                Default is True.
-            _request_timeout (int/float/tuple): timeout setting for this request. If
-                one number provided, it will be total request timeout. It can also
-                be a pair (tuple) of (connection, read) timeouts.
-                Default is None.
-            _check_input_type (bool): specifies if type checking
-                should be done one the data sent to the server.
-                Default is True.
-            _check_return_type (bool): specifies if type checking
-                should be done one the data received from the server.
-                Default is True.
-            _host_index (int/None): specifies the index of the server
-                that we want to use.
-                Default is read from the configuration.
-            async_req (bool): execute request asynchronously
-
-        Returns:
-            None
-                If the method is called asynchronously, returns the request
-                thread.
-        """
-        kwargs['async_req'] = kwargs.get(
-            'async_req', False
-        )
-        kwargs['_return_http_data_only'] = kwargs.get(
-            '_return_http_data_only', True
-        )
-        kwargs['_preload_content'] = kwargs.get(
-            '_preload_content', True
-        )
-        kwargs['_request_timeout'] = kwargs.get(
-            '_request_timeout', None
-        )
-        kwargs['_check_input_type'] = kwargs.get(
-            '_check_input_type', True
-        )
-        kwargs['_check_return_type'] = kwargs.get(
-            '_check_return_type', True
-        )
-        kwargs['_host_index'] = kwargs.get('_host_index')
-        kwargs['repository'] = \
-            repository
-        kwargs['refs_dump'] = \
-            refs_dump
-        return self.restore_refs_endpoint.call_with_http_info(**kwargs)
