+++ clients/python/test/test_refs_api.py
@@ -31,13 +31,6 @@ class TestRefsApi(unittest.TestCase):
         """
         pass
 
-    def test_dump_refs(self):
-        """Test case for dump_refs
-
-        Dump repository refs (tags, commits, branches) to object store  # noqa: E501
-        """
-        pass
-
     def test_find_merge_base(self):
         """Test case for find_merge_base
 
@@ -56,13 +49,6 @@ class TestRefsApi(unittest.TestCase):
         """Test case for merge_into_branch
 
         merge references  # noqa: E501
-        """
-        pass
-
-    def test_restore_refs(self):
-        """Test case for restore_refs
-
-        Restore repository refs (tags, commits, branches) from object store  # noqa: E501
         """
         pass
